+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Next Release (TBD)
 ------------------
 * Stored `ruby_context` from which a `Garner::Cache::Identity` was initialized as an `attr_accessor` on the object.
+* Fixed `cache_enabled?` logic and added a `nocache` declaration to `Garner::Cache::Identity`.
 
 0.4.2 (6/28/2013)
 -----------------